+++ tests/fixtures/external_plugins.py
@@ -26,11 +26,14 @@ logger.info(
 
 
 def function_name(plugin: str ='flake8') ->str:
+    """Test `wrong`rst code."""
     return plugin
 
 
 multiline_string = 'some\
 string'
+
+'\''
 
 print('test')