+++ tests/test_checker/test_noqa.py
@@ -5,10 +5,11 @@ import subprocess
 from collections import Counter
 
 ERROR_PATTERN = re.compile(r'(Z\d{3})')
-IGNORED_VIOLATIONS = [
+IGNORED_VIOLATIONS = (
     'Z202',  # since our test case is complex, that's fine
+    'Z226',  # we have a lot of ugly strings inside,
     'Z402',  # since we obviously use a lot of `noqa` comments
-]
+)
 
 
 def _assert_errors_count_in_output(output, errors, all_violations):
@@ -68,6 +69,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z223': 1,
         'Z224': 1,
         'Z225': 1,
+        'Z226': 0,
 
         'Z300': 1,
         'Z301': 1,
@@ -100,6 +102,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z328': 2,
         'Z329': 1,
         'Z330': 1,
+        'Z331': 1,
 
         'Z400': 0,
         'Z401': 0,
@@ -109,6 +112,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z410': 1,
         'Z411': 0,
         'Z412': 0,
+        'Z413': 1,
         'Z420': 2,
         'Z421': 1,
         'Z422': 1,
@@ -152,6 +156,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z463': 1,
         'Z464': 1,
         'Z465': 1,
+        'Z466': 1,
     }
 
     process = subprocess.Popen(