+++ tests/test_plugins.py
@@ -47,8 +47,10 @@ def test_noqa_fixture_disabled(absolute_path):
     assert output.count('N802') > 0  # pep8-naming
     assert output.count('P101') > 0  # flake8-string-format
     assert output.count('Q000') > 0  # flake8-quotes
+    assert output.count('Q003') > 0  # flake8-quotes
     assert output.count('S001') > 0  # flake8-pep3101
     assert output.count('S101') > 0  # flake8-bandit
     assert output.count('T001') > 0  # flake8-print
     assert output.count('T100') > 0  # flake8-debugger
     assert output.count('TAE002') > 0  # flake8-annotations-complexity
+    assert output.count('RST299') > 0  # flake8-rst-docstrings