+++ wemake_python_styleguide/types.py
@@ -40,7 +40,7 @@ Reference
 import ast
 from typing import Tuple, Type, Union
 
-from typing_extensions import Final, Protocol, final  # noqa: F401
+from typing_extensions import Protocol
 
 #: In cases we need to work with both import types.
 AnyImport = Union[ast.Import, ast.ImportFrom]
@@ -99,3 +99,4 @@ class ConfigurationOptions(Protocol):
     max_imports: int
     max_base_classes: int
     max_decorators: int
+    max_string_usages: int