+++ wemake_python_styleguide/violations/consistency.py
@@ -56,6 +56,7 @@ Summary
    UselessNodeViolation
    UselessExceptCaseViolation
    UselessOperatorsViolation
+   InconsistentReturnVariableViolation
 
 Consistency checks
 ------------------
@@ -91,10 +92,12 @@ Consistency checks
 .. autoclass:: UselessNodeViolation
 .. autoclass:: UselessExceptCaseViolation
 .. autoclass:: UselessOperatorsViolation
+.. autoclass:: InconsistentReturnVariableViolation
 
 """
 
-from wemake_python_styleguide.types import final
+from typing_extensions import final
+
 from wemake_python_styleguide.violations.base import (
     ASTViolation,
     TokenizeViolation,
@@ -548,6 +551,7 @@ class MissingSpaceBetweenKeywordAndParenViolation(TokenizeViolation):
     code = 313
 
 
+@final
 class WrongConditionalViolation(ASTViolation):
     """
     Forbids using ``if`` statements that use invalid conditionals.
@@ -576,6 +580,7 @@ class WrongConditionalViolation(ASTViolation):
     code = 314
 
 
+@final
 class ObjectInBaseClassesListViolation(ASTViolation):
     """
     Forbids extra ``object`` in parent classes list.
@@ -1127,6 +1132,7 @@ class UselessNodeViolation(ASTViolation):
     code = 328
 
 
+@final
 class UselessExceptCaseViolation(ASTViolation):
     """
     Forbids to use meaningless ``except`` cases.
@@ -1163,6 +1169,7 @@ class UselessExceptCaseViolation(ASTViolation):
     code = 329
 
 
+@final
 class UselessOperatorsViolation(ASTViolation):
     """
     Forbids the use of unnecessary operators in your code.
@@ -1198,3 +1205,36 @@ class UselessOperatorsViolation(ASTViolation):
 
     code = 330
     error_template = 'Found unnecessary operator: {0}'
+
+
+@final
+class InconsistentReturnVariableViolation(ASTViolation):
+    """
+    Forbid local variable that are only used in ``return`` statements.
+
+    Reasoning:
+        This is done for consistency and more readable source code.
+
+    Solution:
+        Forbid to use local variables that are only used in `return` statements
+
+    Example::
+
+        # Correct:
+        def some_function():
+            return 1
+
+        # Wrong:
+        def some_function():
+            some_value = 1
+            return some_value
+
+
+    .. versionadded:: 0.9.0
+
+    """
+
+    error_template = (
+        'Found local variable that are only used in `return` statements'
+    )
+    code = 331