+++ wemake_python_styleguide/visitors/ast/classes.py
@@ -4,9 +4,15 @@ import ast
 from collections import Counter
 from typing import ClassVar, FrozenSet, List
 
+from typing_extensions import final
+
 from wemake_python_styleguide import constants, types
 from wemake_python_styleguide.logics.functions import get_all_arguments
-from wemake_python_styleguide.logics.nodes import is_contained, is_doc_string
+from wemake_python_styleguide.logics.nodes import (
+    get_context,
+    is_contained,
+    is_doc_string,
+)
 from wemake_python_styleguide.violations.best_practices import (
     BadMagicMethodViolation,
     BaseExceptionSubclassViolation,
@@ -25,7 +31,7 @@ from wemake_python_styleguide.visitors.base import BaseNodeVisitor
 from wemake_python_styleguide.visitors.decorators import alias
 
 
-@types.final
+@final
 class WrongClassVisitor(BaseNodeVisitor):
     """
     This class is responsible for restricting some ``class`` anti-patterns.
@@ -91,7 +97,7 @@ class WrongClassVisitor(BaseNodeVisitor):
         self.generic_visit(node)
 
 
-@types.final
+@final
 @alias('visit_any_function', (
     'visit_FunctionDef',
     'visit_AsyncFunctionDef',
@@ -130,7 +136,7 @@ class WrongMethodVisitor(BaseNodeVisitor):
                 self.add_violation(StaticMethodViolation(node))
 
     def _check_bound_methods(self, node: types.AnyFunctionDef) -> None:
-        node_context = getattr(node, 'wps_context', None)
+        node_context = get_context(node)
         if not isinstance(node_context, ast.ClassDef):
             return
 
@@ -148,7 +154,7 @@ class WrongMethodVisitor(BaseNodeVisitor):
                 self.add_violation(YieldInsideInitViolation(node))
 
 
-@types.final
+@final
 class WrongSlotsVisitor(BaseNodeVisitor):
     """Visits class attributes."""
 
@@ -166,9 +172,7 @@ class WrongSlotsVisitor(BaseNodeVisitor):
             IncorrectSlotsViolation
 
         """
-        context = getattr(node, 'wps_context', None)
-        if isinstance(context, ast.ClassDef):
-            self._check_slots(node)
+        self._check_slots(node)
         self.generic_visit(node)
 
     def _contains_slots_assign(self, node: ast.Assign) -> bool:
@@ -195,6 +199,10 @@ class WrongSlotsVisitor(BaseNodeVisitor):
                 return
 
     def _check_slots(self, node: ast.Assign) -> None:
+        context = get_context(node)
+        if not isinstance(context, ast.ClassDef):
+            return
+
         if not self._contains_slots_assign(node):
             return