+++ wemake_python_styleguide/visitors/base.py
@@ -66,9 +66,11 @@ import ast
 import tokenize
 from typing import List, Sequence, Type
 
+from typing_extensions import final
+
 from wemake_python_styleguide import constants
 from wemake_python_styleguide.logics.filenames import get_stem
-from wemake_python_styleguide.types import ConfigurationOptions, final
+from wemake_python_styleguide.types import ConfigurationOptions
 from wemake_python_styleguide.violations.base import BaseViolation
 
 
@@ -80,7 +82,7 @@ class BaseVisitor(object):
         options: contains the options objects passed and parsed by ``flake8``.
         filename: filename passed by ``flake8``, each visitor has a file name.
         violations: list of :term:`violations <violation>`
-            for the specific visitor.
+        for the specific visitor.
 
     """