+++ wemake_python_styleguide/visitors/tokenize/statements.py
@@ -4,8 +4,9 @@ import tokenize
 from collections import defaultdict
 from typing import ClassVar, DefaultDict, Dict, List, Sequence, Set, Tuple
 
+from typing_extensions import final
+
 from wemake_python_styleguide.logics.tokens import only_contains
-from wemake_python_styleguide.types import final
 from wemake_python_styleguide.violations.consistency import (
     ExtraIndentationViolation,
     WrongBracketPositionViolation,
@@ -64,10 +65,8 @@ class ExtraIndentationVisitor(BaseTokenVisitor):
 
     def _get_token_offset(self, token: tokenize.TokenInfo) -> int:
         if token.exact_type == tokenize.INDENT:
-            offset = token.end[1]
-        else:
-            offset = token.start[1]
-        return offset
+            return token.end[1]
+        return token.start[1]
 
     def _check_individual_line(
         self,