+++ CHANGELOG.md
@@ -3,16 +3,54 @@
 We follow Semantic Versions since the `0.1.0` release.
 We used to have incremental versioning before `0.1.0`.
 
-## 0.9.0
+
+## WIP
 
 ### Features
 
-- Forbid to use multiline conditions
+- Validates options that are passed with `flake8`
+- Forbids to use module level mutable constants
+- Forbids to over-use strings
+- Forbids to use `breakpoint` function
+- Adds `flake8-rst-docstrings` as a dependency
+
+
+## 0.9.0
+
+This is mostly a supporting release with several new features
+and lots of bug fixes.
+
+### Features
+
+- Forbids to use magic module methods `__getattr__` and `__dir__`
+- Forbids to use multiline conditions
+- Forbids local variables that are only used in `return` statements
 
 ### Bugfixes
 
-- Fixes problem with missing _allowed_left_nodes
-- Fixes problem false positive for Z121 when using `_` for unused var names
+- Fixes module names for modules like `io.py`
+- Fixes false positive `Z310` for numbers like `0xE`
+- Fixes false positive for compare ordering with `await`
+- Fixes problem with missing `_allowed_left_nodes`
+- Fixes problem false positive for `Z121` when using `_` for unused var names
+- Fixes false positive for negative number in default values
+- Fixes error text for `ComplexDefaultValuesViolation`
+- Fixes problem with false positive for `Z459`
+  when a default value is an `Ellipsis`
+
+### Misc
+
+- Adds `py.typed` file in case someone will import our code,
+  now it will have types
+- Adds several missing `@final` decorators
+- Enforces typing support
+- Refactors how `typing_extensions` package is used
+- Adds docs about `black`
+- Adds big "Star" button
+- Multiple dependencies update
+- Better `exclude` rule for `flake8` check
+- Removed warnings from `pytest`
+
 
 ## 0.8.1